% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kread.R
\name{kread}
\alias{kread}
\title{kread - A utility function to read data files based on their extensions}
\usage{
kread(path, ...)
}
\arguments{
\item{path}{\link{character} specifying the file path to be read. The file
must have an extension to determine the appropriate reader function.}

\item{...}{Additional arguments passed to the underlying reader function.}
}
\value{
A data frame or tibble containing the data from the file.
}
\description{
This function reads data files based on their file extensions and returns the
corresponding data frame or tibble. It supports multiple file formats such as
CSV, TSV, Excel (XLS, XLSX), and ODS.
}
\details{
The function determines the file type by extracting the file extension from
the provided \code{path}. It then uses the appropriate reader function from
packages such as \code{readr}, \code{readxl}, or \code{readODS} to read the file. If the
file extension is not recognized or the path does not reference a file, an
error is raised.

Supported file extensions and their corresponding reader functions:
\itemize{
\item \code{csv}: \code{readr::read_delim}
\item \code{tsv}: \code{readr::read_delim}
\item \code{xls}: \code{readxl::read_excel}
\item \code{xlsx}: \code{readxl::read_excel}
\item \code{ods}: \code{readODS::read_ods}
}
}
\examples{
\dontrun{
# Reading a CSV file
data <- kread("data.csv")

# Reading an Excel file
data <- kread("data.xlsx", sheet = 1)

# Reading an ODS file
data <- kread("data.ods")
}

}
