% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-kmerge.R
\name{kmergeListItems}
\alias{kmergeListItems}
\alias{kmergeMultipleLists}
\title{Merge List Items}
\usage{
kmergeListItems(x, ..., cfg = loadConfig())

kmergeMultipleLists(x, y, ..., cfg = loadConfig())
}
\arguments{
\item{x}{A list to be merged.}

\item{...}{Additional arguments passed to the merging functions.}

\item{cfg}{a config object. Configuration to use. See \link{loadConfig} for
details}

\item{y}{An optional second list to merge with \code{x}. If \code{NULL}, merges items
within \code{x}.}
}
\value{
A single merged object from the list items.

Throws an error indicating the function is not implemented.
}
\description{
Merges all items within a list into a single object.

Placeholder function for merging multiple lists. Not implemented
yet.
}
